+++ crates/matrix-sdk-base/src/client.rs
@@ -512,9 +512,11 @@ impl BaseClient {
                     .flatten()
                     .filter_map(|room_key_info| self.get_room(&room_key_info.room_id))
                     .collect(),
-                olm_machine.as_ref(),
-                self.decryption_trust_requirement,
-                self.handle_verification_events,
+                processors::e2ee::E2EE::new(
+                    olm_machine.as_ref(),
+                    self.decryption_trust_requirement,
+                    self.handle_verification_events,
+                ),
             )
             .await?;
