+++ crates/matrix-sdk-base/src/response_processors/latest_event.rs
@@ -13,12 +13,10 @@
 // limitations under the License.
 
 use matrix_sdk_common::deserialized_responses::TimelineEvent;
-use matrix_sdk_crypto::{
-    DecryptionSettings, OlmMachine, RoomEventDecryptionResult, TrustRequirement,
-};
+use matrix_sdk_crypto::{DecryptionSettings, RoomEventDecryptionResult};
 use ruma::{events::AnySyncTimelineEvent, serde::Raw, RoomId};
 
-use super::{verification, Context};
+use super::{e2ee::E2EE, verification, Context};
 use crate::{
     latest_event::{is_suitable_for_latest_event, LatestEvent, PossibleLatestEvent},
     Result, Room,
@@ -33,27 +31,19 @@ use crate::{
 pub async fn decrypt_from_rooms(
     context: &mut Context,
     rooms: Vec<Room>,
-    olm_machine: Option<&OlmMachine>,
-    decryption_trust_requirement: TrustRequirement,
-    verification_is_allowed: bool,
+    e2ee: E2EE<'_>,
 ) -> Result<()> {
-    let Some(olm_machine) = olm_machine else {
+    // All functions used by this one expect an `OlmMachine`. Return if there is
+    // none.
+    if e2ee.olm_machine.is_none() {
         return Ok(());
-    };
+    }
 
     for room in rooms {
         // Try to find a message we can decrypt and is suitable for using as the latest
         // event. If we found one, set it as the latest and delete any older
         // encrypted events
-        if let Some((found, found_index)) = find_suitable_and_decrypt(
-            context,
-            olm_machine,
-            &room,
-            &decryption_trust_requirement,
-            verification_is_allowed,
-        )
-        .await
-        {
+        if let Some((found, found_index)) = find_suitable_and_decrypt(context, &room, &e2ee).await {
             room.on_latest_event_decrypted(
                 found,
                 found_index,
@@ -68,10 +58,8 @@ pub async fn decrypt_from_rooms(
 
 async fn find_suitable_and_decrypt(
     context: &mut Context,
-    olm_machine: &OlmMachine,
     room: &Room,
-    decryption_trust_requirement: &TrustRequirement,
-    verification_is_allowed: bool,
+    e2ee: &E2EE<'_>,
 ) -> Option<(Box<LatestEvent>, usize)> {
     let enc_events = room.latest_encrypted_events();
     let power_levels = room.power_levels().await.ok();
@@ -82,14 +70,8 @@ async fn find_suitable_and_decrypt(
         // Size of the `decrypt_sync_room_event` future should not impact this
         // async fn since it is likely that there aren't even any encrypted
         // events when calling it.
-        let decrypt_sync_room_event = Box::pin(decrypt_sync_room_event(
-            context,
-            olm_machine,
-            event,
-            room.room_id(),
-            decryption_trust_requirement,
-            verification_is_allowed,
-        ));
+        let decrypt_sync_room_event =
+            Box::pin(decrypt_sync_room_event(context, event, &e2ee, room.room_id()));
 
         if let Ok(decrypted) = decrypt_sync_room_event.await {
             // We found an event we can decrypt
@@ -119,19 +101,21 @@ async fn find_suitable_and_decrypt(
 /// representing the decryption error; in the case of problems with our
 /// application, returns `Err`.
 ///
-/// Returns `Ok(None)` if encryption is not configured.
+/// # Panics
+///
+/// Panics if there is no [`OlmMachine`] in [`E2EE`].
 async fn decrypt_sync_room_event(
     context: &mut Context,
-    olm_machine: &OlmMachine,
     event: &Raw<AnySyncTimelineEvent>,
+    e2ee: &E2EE<'_>,
     room_id: &RoomId,
-    decryption_trust_requirement: &TrustRequirement,
-    verification_is_allowed: bool,
 ) -> Result<TimelineEvent> {
     let decryption_settings =
-        DecryptionSettings { sender_device_trust_requirement: *decryption_trust_requirement };
+        DecryptionSettings { sender_device_trust_requirement: e2ee.decryption_trust_requirement };
 
-    let event = match olm_machine
+    let event = match e2ee
+        .olm_machine
+        .expect("An `OlmMachine` is expected")
         .try_decrypt_room_event(event.cast_ref(), room_id, &decryption_settings)
         .await?
     {
@@ -142,8 +126,7 @@ async fn decrypt_sync_room_event(
                 verification::process_if_relevant(
                     context,
                     &sync_timeline_event,
-                    verification_is_allowed,
-                    Some(olm_machine),
+                    e2ee.clone(),
                     room_id,
                 )
                 .await?;
@@ -167,11 +150,8 @@ mod tests {
     };
     use ruma::{event_id, events::room::member::MembershipState, room_id, user_id};
 
-    use super::{decrypt_from_rooms, Context};
-    use crate::{
-        rooms::normal::RoomInfoNotableUpdateReasons, test_utils::logged_in_base_client,
-        StateChanges,
-    };
+    use super::{decrypt_from_rooms, Context, E2EE};
+    use crate::{rooms::normal::RoomInfoNotableUpdateReasons, test_utils::logged_in_base_client};
 
     #[async_test]
     async fn test_when_there_are_no_latest_encrypted_events_decrypting_them_does_nothing() {
@@ -208,9 +188,11 @@ mod tests {
         decrypt_from_rooms(
             &mut context,
             vec![room.clone()],
-            client.olm_machine().await.as_ref(),
-            client.decryption_trust_requirement,
-            client.handle_verification_events,
+            E2EE::new(
+                client.olm_machine().await.as_ref(),
+                client.decryption_trust_requirement,
+                client.handle_verification_events,
+            ),
         )
         .await
         .unwrap();
